requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
google-genai